requirements.txt
selenium
opencv-python
Pillow